requirements.txt
absl-py==2.0.0
accelerate==0.23.0
addict==2.4.0
aiofiles==23.2.1
aiohttp==3.8.5
aiosignal==1.3.1
albumentations==1.3.1
altair==5.1.1
annotated-types==0.5.0
anyio==3.7.1
appdirs==1.4.4
astor==0.8.1
astunparse==1.6.3
async-timeout==4.0.3
asyncer==0.0.2
attrdict==2.0.1
attrs==23.1.0
auto-gptq==0.3.0
autosub==0.3.12
Babel==2.12.1
backgroundremover==0.2.5
bce-python-sdk==0.8.90
beautifulsoup4==4.12.2
bitsandbytes==0.40.2
black==23.7.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
cmake==3.27.5
cnocr==2.2.4
cnstd==1.2.3.4
coloredlogs==15.0.1
commandlines==0.4.1
contourpy==1.1.1
cssselect==1.2.0
cssutils==2.7.1
cycler==0.11.0
Cython==3.0.2
datasets==2.14.5
decorator==4.4.2
deepface==0.0.79
Deprecated==1.2.14
deprecation==2.1.0
diffusers==0.21.4
dill==0.3.7
diskcache==5.6.3
dlib==19.24.2
docker-pycreds==0.4.0
docopt==0.6.2
easydict==1.10
et-xmlfile==1.1.0
evaluate==0.4.0
exceptiongroup==1.1.3
face-recognition==1.3.0
face-recognition-models==0.3.0
fastapi==0.100.0
ffmpeg-python==0.2.0
ffmpy==0.3.1
filelock==3.12.4
filetype==1.2.0
fire==0.5.0
Flask==2.3.3
flask-babel==3.1.0
flatbuffers==23.5.26
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.6.0
future==0.18.3
gast==0.5.4
gdown==4.7.1
gitdb==4.0.10
GitPython==3.1.37
google-api-core==2.12.0
google-api-python-client==2.101.0
google-auth==2.23.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.60.0
gradio==3.37.0
gradio_client==0.5.3
grpcio==1.58.0
gTTS==2.3.2
gunicorn==21.2.0
h11==0.14.0
h5py==3.9.0
handi==0.10.4.1
hsh==1.1.0
httpcore==0.18.0
httplib2==0.22.0
httpx==0.25.0
huggingface-hub==0.16.4
humanfriendly==10.0
HyperPyYAML==1.2.2
idna==3.4
imageio==2.31.4
imageio-ffmpeg==0.4.9
img2pdf==0.4.4
img2txt.py==2.4
imgaug==0.4.0
importlib-metadata==6.8.0
importlib-resources==6.1.0
iniconfig==2.0.0
insightface==0.7.3
invisible-watermark==0.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
keras==2.14.0
kiwisolver==1.4.5
langdetect==1.0.9
lazy_loader==0.3
libclang==16.0.6
lightning-utilities==0.9.0
linkify-it-py==2.0.2
lit==17.0.1
llama-cpp-python==0.1.77
llama2-wrapper==0.1.12
llvmlite==0.41.0
lmdb==1.4.1
lpips==0.1.4
lxml==4.9.3
Markdown==3.4.4
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.8.0
mdit-py-plugins==0.3.3
mdurl==0.1.2
memory-profiler==0.61.0
ml-dtypes==0.2.0
more-itertools==10.1.0
moviepy==1.0.3
mpmath==1.3.0
mtcnn==0.1.1
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
networkx==3.1
ngrok==0.10.0
nltk==3.8.1
numba==0.58.0
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
onnx==1.14.0
onnxruntime==1.15.0
openai-whisper==20230918
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.7.0.72
openpyxl==3.1.2
openvino==2023.1.0
openvino-telemetry==2023.1.1
opt-einsum==3.3.0
optimum==1.13.2
optimum-intel==1.11.0
orjson==3.9.7
packaging==23.1
paddle-bfloat==0.1.7
paddleocr==2.7.0.3
paddlepaddle==2.5.1
panda==0.3.1
pandas==2.1.1
pathspec==0.11.2
pathtools==0.1.2
pdf2docx==0.5.6
peft==0.5.0
pikepdf==8.4.1
Pillow==10.0.1
platformdirs==3.10.0
pluggy==1.3.0
Polygon3==3.0.9.1
pooch==1.7.0
premailer==3.10.0
prettytable==3.9.0
proglog==0.1.10
progressbar==2.5
protobuf==3.20.3
psutil==5.9.5
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyclipper==1.3.0.post5
pycparser==2.21
pycryptodome==3.19.0
pydantic==2.2.1
pydantic-settings==2.0.3
pydantic_core==2.6.1
pydub==0.25.1
Pygments==2.16.1
PyMatting==1.1.8
PyMuPDF==1.20.2
pyparsing==3.1.1
PySocks==1.7.1
pysrt==1.1.2
pytesseract==0.3.10
pytest==7.4.0
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-multipart==0.0.6
pytorch-lightning==2.0.9.post0
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
qudida==0.0.4
rapidfuzz==3.3.1
rarfile==4.1
referencing==0.30.2
regex==2023.8.8
rembg==2.0.50
requests==2.31.0
requests-oauthlib==1.3.1
responses==0.18.0
retina-face==0.0.13
rouge==1.0.1
rpds-py==0.10.3
rsa==4.9
ruamel.yaml==0.17.33
ruamel.yaml.clib==0.2.7
sacremoses==0.0.53
safetensors==0.3.3
scikit-image==0.20.0
scikit-learn==1.3.1
scipy==1.11.1
seaborn==0.13.0
semantic-version==2.10.0
sentencepiece==0.1.99
sentry-sdk==1.31.0
setproctitle==1.3.2
shapely==2.0.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
speechbrain==0.5.15
SpeechRecognition==3.10.0
srt==3.5.3
sse-starlette==1.6.5
starlette==0.27.0
sympy==1.12
tb-nightly==2.15.0a20230925
tensorboard==2.14.1
tensorboard-data-server==0.7.1
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
tesserocr==2.6.1
textblob==0.17.1
threadpoolctl==3.2.0
tifffile==2023.9.18
tiktoken==0.3.3
tokenizers==0.13.3
tomli==2.0.1
toolz==0.12.0
torch==2.0.1
torchaudio==2.0.2
torchmetrics==1.2.0
torchvision==0.15.2
tqdm==4.65.0
transformers==4.33.3
triton==2.0.0
typing_extensions==4.8.0
tzdata==2023.3
uc-micro-py==1.0.2
Unidecode==1.3.7
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.23.1
video2audio==2023.1.14
videogrep==2.1.3
visualdl==2.5.3
vosk==0.3.45
waitress==2.1.2
wandb==0.15.11
watchdog==3.0.0
wcwidth==0.2.6
websockets==11.0.3
Werkzeug==2.3.7
wrapt==1.14.1
xxhash==3.3.0
yapf==0.40.2
yarl==1.9.2
zipp==3.17.0